feat: create the user management view
refactor: use Link from react-router in Breadcrumb links

feat: create RoleCard component

This is a reusable component, that display a card for each role with a collapsible showing the associated permissions. The permissions, are organized by  resource and enable/disable.

feat: integrate the LibrariesUserManagement view in the application

style: add ts doc to buildPermissionsByRoleMatrix  function